Changed the "network" files
I think that I made most of the changes that are required for the
"network" files. I have also editted the lsp_client files to what I
believe they should be. The lsp_server file still needs changed to RPC.
I created the interface so someone needs to simply implement the RPC
function send_message to read the messages that the clients send and
just return a message back.