Finally! make -f Makefile.linux server compiles: see description
After tons of research and hours, I've finally managed to figure out
why the RPC stuff wouldn't compile and how to transition compile it's
stuff first in CC and then the rest of the program in g++. For Mac OS
and Linux, the compiler needed to be CC but we still needed g++ since
we were using C++ stuff. So you all can ignore the folder Cracker
Tester and its contents unless you want to test the barebones RPC code.
Otherwise, All the components for the server now compile using make -f
Makefile.linux server. All the RPC code and the network code and the
lsp_server code compile. Server.cpp doesn't yet, but that's because
changes need to be made there. The client side also needs to be
implemented, but that should be easy to test now that I know how to
compile all this stuff. Let me know if you still have any further
questions